requirements.txt
discord.py[voice]>=2.3.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
aiosqlite>=0.19.0
pytz>=2023.3